Added tests

+++ ClownsProject/Controllers/UserController.cs
@@ -15,25 +15,28 @@ namespace ClownsProject.Controllers
     {
         public static User CurrentUser { get; set; }
 
-        public static void Delete(string login)
+        public static bool Delete(string login)
         {
             using (var db = new MortalkombatContext())
             {
-                var user = db.Users.First(u => u.Login == login);
+                var user = db.Users.FirstOrDefault(u => u.Login == login);
+                if (user == null)
+                    return false;
                 var tasks = TaskService.GetTasks().Where(t => t.Login == login && t.IdStatus != 3).ToList();
                 if (tasks.Count != 0)
                 {
                     MessageBox.Show("Пока у сотрудника есть незавершённые задачи его нельзя уволить!");
-                    return;
+                    return false;
                 }
                 tasks = TaskService.GetTasks().Where(t => t.Login == login && t.IdStatus == 3).ToList();
                 db.Tasks.RemoveRange(tasks);
                 db.Users.Remove(user);
                 db.SaveChanges();
                 MessageBox.Show("Увольнение произошло успешно!");
+                return true;
             }
         }
-        public static void Authorization(string login, string pass)
+        public static bool Authorization(string login, string pass)
         {
             using (var db = new MortalkombatContext())
             {
@@ -42,28 +45,34 @@ namespace ClownsProject.Controllers
                 if (CurrentUser != null)
                 {
                     MessageBox.Show("Авторизация прошла успешно!");
-                    if (CurrentUser.IdRoleNavigation.Title.Equals("администратор"))
+                    try
                     {
-                        PageController.ShowAdminPage();
-                    }
-                    else if (CurrentUser.IdRoleNavigation.Title.Equals("менеджер"))
-                    {
-                        PageController.ShowProjectsPage();
-                    }
-                    else
-                    {
-                        var tasks = TaskService.GetTasks();
+                        if (CurrentUser.IdRoleNavigation.Title.Equals("администратор"))
+                        {
+                            PageController.ShowAdminPage();
+                        }
+                        else if (CurrentUser.IdRoleNavigation.Title.Equals("менеджер"))
+                        {
+                            PageController.ShowProjectsPage();
+                        }
+                        else
+                        {
+                            var tasks = TaskService.GetTasks();
 
-                        PageController.ShowTasksPage(tasks);
-                    }
+                            PageController.ShowTasksPage(tasks);
+                        }
+                    } catch(Exception ex) { }
+                    return true;
+                    
                 }
                 else
                 {
                     MessageBox.Show("Пользователь не найден!");
+                    return false;
                 }
             }
         }
-        public static void Registration(string login, string password, string phoneNumber)
+        public static bool Registration(string login, string password, string phoneNumber)
         {
             var phoneRegex_1 = new Regex(@"\+7\([0-9][0-9][0-9]\)[0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]");
             var phoneRegex_2 = new Regex(@"8\([0-9][0-9][0-9]\)[0-9][0-9][0-9]-[0-9][0-9]-[0-9][0-9]");
@@ -83,22 +92,29 @@ namespace ClownsProject.Controllers
                             Passwords = password,
                             PhoneNumber = phoneNumber,
                             IdRole = 1,
-                            Brand = "none",
+                            Brand = "test",
                         };
                         db.Users.Add(user);
                         db.SaveChanges();
                         MessageBox.Show("Регистрация прошла успешно!");
-                        PageController.ShowAdminPage();
+                        try
+                        {
+                            PageController.ShowAdminPage();
+                        }
+                        catch (Exception ex) { }
+                        return true;
                     }
                     else
                     {
                         MessageBox.Show("Такой пользователь уже существует!");
+                        return false;
                     }
                 }
             }
             else
             {
                 MessageBox.Show("Неверый формат номера телефона!\nНомер должен соответсвовать одному из четырёх форматов:\n+7(XXX)XXX-XX-XX\n8(XXX)XXX-XX-XX\n+7XXXXXXXXXX\n8XXXXXXXXXX");
+                return false;
             }
         }
